requirements.txt
amqp==2.6.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.24.55
botocore==1.27.55
celery==4.4.6
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1
django-environ==0.9.0
django-storages==1.13.1
djangorestframework==3.13.1
drf-yasg==1.21.3
future==0.18.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==4.6.11
MarkupSafe==2.1.1
numpy==1.23.2
packaging==21.3
pandas==1.4.3
prompt-toolkit==3.0.30
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.11
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.1
amqp==2.6.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.24.55
botocore==1.27.55
celery==4.4.6
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1
django-environ==0.9.0
django-storages==1.13.1
djangorestframework==3.13.1
drf-yasg==1.21.3
future==0.18.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==4.6.11
MarkupSafe==2.1.1
numpy==1.23.2
packaging==21.3
pandas==1.4.3
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.11
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.1
